refactor: avoid duplicate code through utilities
