+++ src/pages/holidays/components/HolidaysTable/index.tsx
@@ -24,7 +24,7 @@ import { mockRequirements } from "@mocks/requirements/requirementsTable.mock";
 import { Tooltip } from "@components/overlay/Tooltip";
 import { InfoModal } from "@components/modals/InfoModal";
 import { spacing } from "@design/tokens/spacing";
-import { contractTypeLabels } from "@mocks/contracts/enums";
+import { transformContractValue } from "@utils/text";
 
 import { IHolidaysTable, HolidayTableDataDetails } from "./types";
 import { StyledTd, StyledTh, TooltipWrapper } from "./styles";
@@ -85,23 +85,6 @@ function HolidaysTable(props: HolidaysTableProps) {
   } = usePagination(data);
 
   const displayData = isMobile ? data : currentData;
-
-  const transformContractValue = (contractValue: string): string => {
-    if (!contractValue) return contractValue;
-
-    const contractTypeKey = Object.keys(contractTypeLabels).find((key) =>
-      contractValue.includes(key),
-    );
-
-    if (contractTypeKey) {
-      return contractValue.replace(
-        contractTypeKey,
-        contractTypeLabels[contractTypeKey],
-      );
-    }
-
-    return contractValue;
-  };
 
   const determineVisibleHeaders = () => {
     if (isMobile) {
